+++ src/plone/importexport/browser/configure.zcml
@@ -13,14 +13,14 @@
     <browser:page
         for="Products.CMFCore.interfaces.IFolderish"
         name="import-export"
-        class=".view.ImportExportView"
+        class=".importexport.ImportExportView"
         permission="cmf.ModifyPortalContent"
         />
 
     <browser:page
         for="Products.CMFCore.interfaces.IFolderish"
         name="export"
-        class=".view.ImportExportView"
+        class=".importexport.ImportExportView"
         attribute="export"
         permission="cmf.ModifyPortalContent"
         />
@@ -28,7 +28,7 @@
     <browser:page
         for="Products.CMFCore.interfaces.IFolderish"
         name="imports"
-        class=".view.ImportExportView"
+        class=".importexport.ImportExportView"
         attribute="imports"
         permission="cmf.ModifyPortalContent"
         />
