+++ src/plone/importexport/browser/importexport.py
@@ -20,175 +20,10 @@ import StringIO, cStringIO
 import zipfile
 import fnmatch
 import operator
+import utils
 
 # TODO: in advanced tab, allow user to change this
 EXCLUDED_ATTRIBUTES = ['member', 'parent', 'items', 'changeNote', '@id', 'UID', 'scales', 'items_total', 'table_of_contents', ]
-
-class InMemoryZip(object):
-    def __init__(self):
-        # Create the in-memory file-like object
-        self.in_memory_zip = StringIO.StringIO()
-
-    def append(self, filename_in_zip, file_contents):
-        '''Appends a file with name filename_in_zip and contents of
-        file_contents to the in-memory zip.'''
-        # Get a handle to the in-memory zip in append mode
-        zf = zipfile.ZipFile(self.in_memory_zip, "a", zipfile.ZIP_DEFLATED, False)
-
-        # Write the file to the in-memory zip
-        zf.writestr(filename_in_zip, file_contents)
-
-        # Mark the files as having been created on Windows so that
-        # Unix permissions are not inferred as 0000
-        for zfile in zf.filelist:
-            zfile.create_system = 0
-
-        return self
-
-    def read(self):
-        '''Returns a string with the contents of the in-memory zip.'''
-        self.in_memory_zip.seek(0)
-        return self.in_memory_zip.read()
-
-    def getfiles(self,zip_file):
-        data = {}
-        '''The problem in the standard zipfile module is that when passed a file object (not a filename), it uses that same passed-in file object for every call to the open method. This means that tell and seek are getting called on the same file and so trying to open multiple files within the zip file is causing the file position to be shared and so multiple open calls result in them stepping all over each other. In contrast, when passed a filename, open opens a new file object.
-        error: *** BadZipfile: Bad CRC-32 for file 'Plone.csv'
-        There I realized if I do a seek(0) on the file object before initializing the ZipFile, the error goes away. Don't see why, as I did nothing before on the file object :/'''
-        zip_file.seek(0)
-        zfile = zipfile.ZipFile(zip_file,'r')
-        for name in zfile.namelist():
-            '''.open() returns a file-like object while .read() return a string like object
-            And csv.DictWriter needs a file like object'''
-            # FIXME .open() should work. May be after retriving data from /tmp in the server may solve this HACK
-            data[name] = cStringIO.StringIO()
-            data[name].write(zfile.read(name))
-            data[name].seek(0)
-        return data
-
-class Pipeline(object):
-    # return unique keys from list
-    def getcsvheaders(self,data):
-        # HACK to keep these fields at first in csv
-        header = {'@type':3,'path':2, 'id':1}
-        for dict_ in data:
-            for key in dict_.keys():
-                if key not in header.keys():
-                    header[key]=1
-                else:
-                    header[key] += 1
-
-        result = []
-        header = sorted(header.items(),key=operator.itemgetter(1), reverse=True)
-        for key in header:
-            result.append(key[0])
-        # pdb.set_trace()
-        return result
-
-    def convertjson(self,obj,data_list):
-        csv_output = cStringIO.StringIO()
-
-        url = obj.request.URL
-        id_ = urlparse(url).path.split('/')[1]
-
-        csv_headers =self.getcsvheaders(data_list)
-
-        if not csv_headers:
-            raise BadRequest("check json data, no keys found")
-
-        try:
-            '''The optional restval parameter specifies the value to be written if the dictionary is missing a key in fieldnames. If the dictionary passed to the writerow() method contains a key not found in fieldnames, the optional extrasaction parameter indicates what action to take. If it is set to 'raise' a ValueError is raised. If it is set to 'ignore', extra values in the dictionary are ignored.'''
-            writer = csv.DictWriter(csv_output, fieldnames=csv_headers,restval='Field NA', extrasaction='raise', dialect='excel')
-            writer.writeheader()
-            for data in data_list:
-                for key in data.keys():
-                    if not data[key]:
-                        data[key]="Null"
-                    if isinstance(data[key],(dict,list)):
-
-                        # store blob content and replace url with path
-                        if isinstance(data[key],dict) and 'download' in data[key].keys():
-                            # pdb.set_trace()
-
-                            file_path = data['path']
-                            relative_filepath = '/'.join(file_path.split('/')[1:])
-
-                            try:
-                                if data[key]['content-type'].split('/')[0]=='image':
-                                    file_data = obj.context.restrictedTraverse(str(relative_filepath)+'/image').data
-                                else:
-                                    file_data = obj.context.restrictedTraverse(str(relative_filepath)+'/file').data
-                            except:
-                                print 'Blob data fetching error'
-                            else:
-                                filename = data[key]['filename']
-                                # pdb.set_trace()
-                                data[key]['download'] = file_path+'/'+filename
-                                obj.zip.append(data[key]['download'],file_data)
-
-                        # store html files and replace key[data] with key[download], value= path in zip
-                        elif isinstance(data[key],dict) and 'data' in data[key].keys():
-                            # pdb.set_trace()
-
-                            file_path = data['path']
-
-                            try:
-                                # pdb.set_trace()
-                                if data[key]['content-type'].split('/')[1]=='html':
-                                    file_data = data[key]['data'].encode(data[key]['encoding'])
-                                    del data[key]['data']
-                            except:
-                                print 'html data fetching error'
-                            else:
-                                filename = file_path.split('/')[-1]+'.html'
-                                data[key]['download'] = file_path+'/'+filename
-                                obj.zip.append(data[key]['download'],file_data)
-
-                        # converting list and dict to quoted json
-                        data[key] = json.dumps(data[key])
-
-                writer.writerow(data)
-        except IOError as (errno, strerror):
-                print("I/O error({0}): {1}".format(errno, strerror))
-        else:
-            obj.zip.append(id_+'.csv',csv_output.getvalue())
-            csv_output.close()
-
-
-        return
-
-    def converttojson(self,data):
-        reader = csv.DictReader(data)
-        data = []
-        for row in reader:
-            data.append(row)
-        # jsonify quoted json values
-        data = self.jsonify(data)
-        return data
-
-    # jsonify quoted json values
-    def jsonify(self,data):
-        if isinstance(data,dict):
-            for key in data.keys():
-                data[key] = self.jsonify(data[key])
-        elif isinstance(data,list):
-            for index in range(len(data)):
-                data[index] = self.jsonify(data[index])
-        try:
-            data = json.loads(data)
-        except:
-            pass
-        finally:
-            return data
-
-    def filter(self,data):
-        if isinstance(data,list):
-            for index in range(len(data)):
-                self.filter(data[index])
-        elif isinstance(data,dict):
-            for key in data.keys():
-                if data[key]=="Field NA" or data[key]=="Null":
-                    del data[key]
 
 class ImportExportView(BrowserView):
     """Import/Export page."""
@@ -221,8 +56,11 @@ class ImportExportView(BrowserView):
         if 'items' in data.keys():
             path = []
             for id_ in data['items']:
-                # HACK restapi path> /Plone/folder while zipfile> Plone/folder
-                path.append('/'.join(urlparse(id_['@id']).path.split('/')[1:]))
+                url_path = urlparse(id_['@id']).path
+                if url_path.startswith('/'):
+                    # restapi path> /Plone/folder while zipfile> Plone/folder
+                    url_path = url_path[1:]
+                path.append(url_path)
 
         # del EXCLUDED_ATTRIBUTES from data
         self.exclude_attributes(data)
@@ -306,21 +144,20 @@ class ImportExportView(BrowserView):
     def export(self):
 
         # create zip in memory
-        self.zip = InMemoryZip()
+        self.zip = utils.InMemoryZip()
 
         # defines Pipeline
-        self.conversion = Pipeline()
+        self.conversion = utils.Pipeline()
 
         if self.request.method == 'POST':
 
-            # get home_path of Plone sites
+            # get id_ of Plone sites
             url = self.request.URL
             id_ = urlparse(url).path.split('/')[1]
-            home_path = '/' + id_
 
             # pdb.set_trace()
             # results is a list of dicts
-            results = self.serialize(self.context, home_path)
+            results = self.serialize(self.context, id_)
 
             self.conversion.convertjson(self,results)
 
@@ -333,6 +170,7 @@ class ImportExportView(BrowserView):
         return
 
     def getparentcontext(self,data):
+        # FIXME for windows too, path>> Plone\site
         path_ = data['path'].split('/')
 
         obj = self.context
@@ -346,10 +184,10 @@ class ImportExportView(BrowserView):
     def imports(self):
 
         # create zip in memory
-        self.zip = InMemoryZip()
+        self.zip = utils.InMemoryZip()
 
         # defines Pipeline
-        self.conversion = Pipeline()
+        self.conversion = utils.Pipeline()
 
         error_log = ''
 
@@ -386,36 +224,41 @@ class ImportExportView(BrowserView):
             for index in range(len(data)):
                 obj_data = data[index]
 
+                # TODO raise the error into log_file
                 if not obj_data['path']:
                     raise BadRequest("Property 'path' is required")
 
                 # pdb.set_trace()
 
                 # FIXME: solution for more than one image/file in an object
-                if 'image' in obj_data.keys():
-                    if obj_data['image']['download'] in files.keys():
+                if obj_data.get('image',None):
+                    value = obj_data['image'].get('download',None)
+                    if value and files.get(value,None):
                         try:
-                            content = files[obj_data['image']['download']].read()
+                            content = files[value].read()
                             obj_data['image']['data'] = content.encode("base64")
                             obj_data['image']['encoding'] = "base64"
                         except:
                             error_log += 'Error in fetching/encoding blob from zip {}'.format(obj_data['path'])
 
-                if 'file' in obj_data.keys():
+                if obj_data.get('file',None):
                     # pdb.set_trace()
-                    if obj_data['file']['download'] in files.keys():
+                    value = obj_data['file'].get('download',None)
+                    if value and files.get(value,None):
                         try:
-                            content = files[obj_data['file']['download']].read()
+                            content = files[value].read()
                             obj_data['file']['data'] = content.encode("base64")
                             obj_data['file']['encoding'] = "base64"
                         except:
                             error_log += 'Error in fetching/encoding blob from zip {}'.format(obj_data['path'])
 
-                if 'text' in obj_data.keys() and obj_data['text']['content-type'].split('/')[-1]=="html":
-                    if obj_data['text']['download'] in files.keys():
+                if obj_data.get('text',None) and obj_data['text'].get('content-type',None):
+                    type_ = obj_data['text']['content-type'].split('/')[-1]
+                    value = obj_data['text'].get('download',None)
+                    if type_=="html" and value and files.get(value,None):
                         try:
                             # pdb.set_trace()
-                            obj_data['text']['data'] = files[obj_data['text']['download']].read()
+                            obj_data['text']['data'] = files[value].read()
                             # obj_data['text']['data'] = content.encode(obj_data['text']['encoding'])
                             del obj_data['text']['download']
                         except:
